Refactor: Turret to module; genericise Building.update()
