+++ src/building.py
@@ -40,13 +40,13 @@ class Building(GameObject):
         for i in range(10, self.SIZE[0] - 10, 20):
             pg.draw.rect(self.image, (200, 200, 200), (i, 10, 10, 10))  # Windows
 
-    def update(self, particles: pg.sprite.Group[Particle]) -> None:
+    def update(self, particles: pg.sprite.Group[Particle], *args, **kwargs) -> None:
         if self.construction_progress < self.CONSTRUCTION_TIME:
             self.construction_progress += 1
             self.image.set_alpha(
                 int(255 * self.construction_progress / self.CONSTRUCTION_TIME)
             )
-        super().update()
+        super().update(*args, **kwargs)
         if self.health <= 0:
             for _ in range(15):
                 particles.add(
